Wordle state was missing, and started with documentation.
