+++ src/games/mod.rs
@@ -1,2 +1,3 @@
 pub mod bee;
 pub mod word;
+pub mod word_state;
